New working examples with renderingOptions
