+++ addon/views/modal.js
@@ -37,7 +37,6 @@ export default Em.View.extend(
   setup: function() {
     this.show();
     this.autofocus();
-    console.log(this.$());
   }.on('didInsertElement'),
 
   show: function() {
