+++ tests/dummy/app/templates/application.hbs
@@ -8,3 +8,5 @@
 {{outlet}}
 
 {{outlet 'modal'}}
+
+{{outlet 'modal-two'}}
